fix: limit seeds instrument edit ,add current team limit , ui 、time and errors count update

+++ src/Services/Masa.Tsc.Service.Admin/Application/Instruments/InstrumentCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
+using static Grpc.Core.Metadata;
+
 namespace Masa.Tsc.Service.Admin.Application.Instruments;
 
 public class InstrumentCommandHandler
@@ -39,6 +41,8 @@ public class InstrumentCommandHandler
         var entry = await _instrumentRepository.GetAsync(command.Data.Id, command.UserId);
         if (entry == null)
             throw new UserFriendlyException($"instrument {command.Data.Id} is not exists");
+        if (!entry.EnableEdit)
+            throw new UserFriendlyException(errorCode: ErrorCodes.NOT_ALLOW_EDIT);
         entry.Update(command.Data);
         await _instrumentRepository.UpdateAsync(entry);
     }
@@ -62,6 +66,10 @@ public class InstrumentCommandHandler
         var entry = await _instrumentRepository.GetDetailAsync(command.InstumentId, command.UserId);
         if (entry == null)
             throw new UserFriendlyException($"instrument {command.InstumentId} is not exists");
+
+        if (!entry.EnableEdit)
+            throw new UserFriendlyException(errorCode: ErrorCodes.NOT_ALLOW_EDIT);
+
         entry.UpdatePanels(command.Data);
         await _instrumentRepository.UpdateDetailAsync(entry);
     }
@@ -75,6 +83,10 @@ public class InstrumentCommandHandler
         var list = await _instrumentRepository.GetListAsync(command.InstrumentIds, command.UserId);
         if (list == null || !list.Any())
             throw new UserFriendlyException("数据不存在");
+
+        if (list.Any(item => !item.EnableEdit))
+            throw new UserFriendlyException(errorCode: ErrorCodes.CONTAINS_NOT_ALLOW_EDIT);
+
         await _instrumentRepository.RemoveRangeAsync(list);
     }
     #endregion
